requirements.txt
numpy==2.3.3
opencv_python==4.12.0.88
Pillow==11.3.0
torch==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
transformers==4.54.0
